Custom Masking (#224)
* Introduce preliminary support for custom masking

* Simplify custom masks to use existing mask property of layers

It makes for a less invasive API if pulley can simply re-use the
existing custom mask applied to it's drawer view. This avoid the
need to add another function to the `PulleyDrawerViewControllerDelegate`
while still allowing us to support a somewhat advanced feature.

* Undo automatic whitespace correction to match current state of public repo

* Consolidate the border mask logic into a single location

* Continuously update mask as the pulley drawer is manipulated

This also gives the host-application a change to update it's custom
mask in `viewWillLayoutSubviews`.

* Account for vertical flip when using the left side display mode

* Add a custom mask example to the demo app

The mask example defaults to off, and simply applies two cutout
"ears" to the top of the panel. This example was chosen because
by not conflicting with views in the Storyboard, it required no
further changes to an otherwise clean demo app.